Refactor the basic admin page into smaller files.
